refactor(mark): more generic print page callback
Rather than assuming that all users of this component has `updateTally` and `resetBallot` steps, just call back once the print job has started and let the parent component handle it.
