+++ apps/mark/frontend/src/pages/print_page.tsx
@@ -1,7 +1,8 @@
-import { useContext } from 'react';
+import { useContext, useEffect, useRef } from 'react';
 import { PrintPage as MarkFlowPrintPage } from '@votingworks/mark-flow-ui';
 import { assert } from '@votingworks/basics';
 import { BallotContext } from '../contexts/ballot_context';
+import { BALLOT_PRINTING_TIMEOUT_SECONDS } from '../config/globals';
 
 export function PrintPage(): JSX.Element {
   const {
@@ -18,6 +19,22 @@ export function PrintPage(): JSX.Element {
   assert(typeof ballotStyleId === 'string', 'ballotStyleId is not defined');
   assert(typeof precinctId === 'string', 'precinctId is not defined');
 
+  const printerTimer = useRef(0);
+
+  function onPrintStarted() {
+    updateTally();
+    printerTimer.current = window.setTimeout(() => {
+      resetBallot(true);
+    }, BALLOT_PRINTING_TIMEOUT_SECONDS * 1000);
+  }
+
+  // Make sure we clean up any pending timeout on unmount
+  useEffect(() => {
+    return () => {
+      clearTimeout(printerTimer.current);
+    };
+  }, []);
+
   return (
     <MarkFlowPrintPage
       electionDefinition={electionDefinition}
@@ -26,8 +43,7 @@ export function PrintPage(): JSX.Element {
       isLiveMode={isLiveMode}
       votes={votes}
       generateBallotId={generateBallotId}
-      updateTally={updateTally}
-      resetBallot={resetBallot}
+      onPrintStarted={onPrintStarted}
     />
   );
 }
