+++ libs/mark-flow-ui/src/config/globals.ts
@@ -1,6 +1,5 @@
 export const IDLE_TIMEOUT_SECONDS = 5 * 60; // VVSG Requirement: 2–5 minutes
 export const IDLE_RESET_TIMEOUT_SECONDS = 45; // VVSG Requirement: 20–45 seconds
-export const BALLOT_PRINTING_TIMEOUT_SECONDS = 5;
 export const WRITE_IN_CANDIDATE_MAX_LENGTH = 40;
 export enum Paths {
   DISPLAY_SETTINGS = '/display-settings',
